+++ aplikasi/ann/ann_models.py
@@ -7,18 +7,6 @@ from aplikasi.ann import (
 from datetime import date, datetime
 from aplikasi.sapi.sapi_models import Sapi
 
-class Prediksi(db.Model):
-    __tablename__ = 'prediksi'
-    id = db.Column(db.Integer(), primary_key=True)
-    sapi_id = db.Column(db.Integer(), db.ForeignKey('sapi.id'))
-    sapi = db.relationship('Sapi')
-    estrus = db.Column(db.Integer())
-    created_at = db.Column(db.DateTime, default=datetime.utcnow)
-    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
-    
-    def __repr__(self):
-        return 'Sapi Estrus : {}'.format(self.sapi_id)
-
 def getDataset():
         df = pd.read_sql_table('dataset', db.engine)
         return df
