+++ aplikasi/templates/ann/prediksi.html
@@ -8,11 +8,7 @@
             <table class="table">
                 <thead>
                   <tr>
-                    <th scope="col">rpf</th>
-                    <th scope="col">perilaku</th>
-                    <th scope="col">ib_ke</th>
-                    <th scope="col">jarak_ib</th>
-                    <th scope="col">laktasi</th>
+                    <th scope="col">sapi_id</th>
                     <th scope="col">prediksi</th>
                   </tr>
                 </thead>
@@ -20,12 +16,8 @@
                 {% for data in prediksi %}
                   <tr>
                     <td>{{ data[0] }}</td>
-                    <td>{{ data[1] }}</td>
-                    <td>{{ data[2] }}</td>
-                    <td>{{ data[3] }}</td>
-                    <td>{{ data[4] }}</td>
                     <td>
-                        {% if data[5]==1 %} Birahi {% else %} Tidak Birahi {% endif %}
+                        {% if data[1]==1 %} Birahi {% else %} Tidak Birahi {% endif %}
                     </td>
                   </tr>
                 {% endfor %}
